moved themed font components
- added custom font options
