+++ package.json
@@ -11,6 +11,8 @@
     "codegen-watch": "graphql-codegen -w"
   },
   "dependencies": {
+    "@expo-google-fonts/roboto": "^0.2.3",
+    "@expo-google-fonts/rubik": "^0.2.3",
     "@react-native-community/datetimepicker": "6.5.2",
     "@react-navigation/drawer": "^6.6.0",
     "@react-navigation/material-top-tabs": "^6.5.3",
@@ -18,15 +20,16 @@
     "@react-navigation/native-stack": "^6.9.8",
     "@tanstack/react-query": "^4.28.0",
     "date-fns": "^2.29.3",
-    "expo": "~47.0.12",
+    "expo": "~47.0.14",
     "expo-application": "~5.0.1",
     "expo-dev-client": "~2.0.1",
+    "expo-font": "~11.0.1",
     "expo-splash-screen": "~0.17.5",
     "expo-status-bar": "~1.4.2",
     "graphql": "^15.8.0",
     "graphql-request": "^5.2.0",
     "react": "18.1.0",
-    "react-native": "0.70.5",
+    "react-native": "0.70.8",
     "react-native-dotenv": "^3.4.7",
     "react-native-element-dropdown": "^2.8.0",
     "react-native-gesture-handler": "~2.8.0",
