+++ src/Shared/DetailsCarousel/CarouselItem.tsx
@@ -2,8 +2,8 @@ import { Pressable, StyleSheet, View } from "react-native";
 import { useTheme } from "@react-navigation/native";
 
 import PlaceholderImage from "../PlaceholderImage";
-import { MainText } from "../ThemedNativeElements";
 import { CarouselEmptyTextProps, CarouselItemProps } from "./types";
+import { MainText } from "../Text";
 
 function CarouselItem(props: CarouselItemProps) {
